Array of Functions

+++ 0x0F-function_pointers/1-array_iterator.c
@@ -1,7 +1,7 @@
 #include "function_pointers.h"
 
 /**
- * array_iterator - an array of function that executes every function 
+ * array_iterator - an array of function that executes every function
  * whenever called
  *
  * @array: an array of integers
@@ -11,7 +11,7 @@
  */
 void array_iterator(int *array, size_t size, void (*action)(int))
 {
-	int i;
+	size_t i;
 
 	for (i = 0; i < size; i++)
 	{
